SI resolution now uses contact id

+++ include/ppx/collision/resolution/sequential_impulses_resolution2D.hpp
@@ -10,7 +10,7 @@ class sequential_impulses_resolution2D : public collision_resolution2D
 {
   public:
     using contact_map =
-        std::unordered_map<kit::non_commutative_tuple<const collider2D *, const collider2D *, std::size_t>,
+        std::unordered_map<kit::non_commutative_tuple<const collider2D *, const collider2D *, std::uint32_t>,
                            contact_constraint2D>;
     using collision_resolution2D::collision_resolution2D;
